+++ hat/assets/js/apps/Iaso/domains/entities/constants.ts
@@ -5,6 +5,14 @@ export const ALGORITHM_DROPDOWN = [
 ];
 
 export const LEVENSHTEIN_PARAMETERS_DROPDOWN = [
-    { label: 'levenshtein_max_distance', value: 'levenshtein_max_distance' },
-    { label: 'above_score_display', value: 'above_score_display' },
+    {
+        label: 'levenshtein_max_distance',
+        value: 'levenshtein_max_distance',
+        value_type: 'number',
+    },
+    {
+        label: 'above_score_display',
+        value: 'above_score_display',
+        value_type: 'number',
+    },
 ];
