+++ hat/assets/js/apps/Iaso/domains/entities/duplicates/list/AnalysisModal.tsx
@@ -3,7 +3,6 @@ import React, {
     FunctionComponent,
     useCallback,
     useEffect,
-    useMemo,
     useState,
 } from 'react';
 import { Grid } from '@mui/material';
@@ -12,10 +11,7 @@ import { AnalysisModalButton } from './AnalysisModalButton';
 import InputComponent from '../../../../components/forms/InputComponent';
 import { useGetBeneficiaryTypesDropdown } from '../../hooks/requests';
 import { useStartAnalyse } from '../hooks/api/analyzes';
-import {
-    ALGORITHM_DROPDOWN,
-    LEVENSHTEIN_PARAMETERS_DROPDOWN,
-} from '../../constants';
+import { ALGORITHM_DROPDOWN } from '../../constants';
 import { formatLabel } from '../../../instances/utils';
 import { useGetFormForEntityType } from '../../entityTypes/hooks/requests/forms';
 import AnalysisModalParameters from './AnalysisModalParameters';
@@ -31,13 +27,11 @@ const AnalysisModal: FunctionComponent<Props> = ({ closeDialog, isOpen }) => {
     const [entityTypeFields, setEntityTypeFields] = useState([]);
     const [referenceForm, setReferenceForm] = useState(undefined);
     const [confirm, setConfirm] = useState(false);
-    const [parameterComponents, setParameterComponents] = useState<string[]>(
-        [],
-    );
+
     const [parameters, setParameters] = useState<
         { name: string; value: string | number }[]
     >([]);
-    const [paramDisabled, setParamDisabled] = useState<boolean[]>([]);
+
     const { data: entityTypesDropdown, isFetching: isFetchingEntityTypes } =
         useGetBeneficiaryTypesDropdown();
 
@@ -80,57 +74,6 @@ const AnalysisModal: FunctionComponent<Props> = ({ closeDialog, isOpen }) => {
         }
     };
 
-    const handleParametersChange = (keyValue, value, index) => {
-        setParameters(prevParams => {
-            const updatedParams = [...prevParams];
-
-            if (keyValue === 'parameters') {
-                const selectedParameter = LEVENSHTEIN_PARAMETERS_DROPDOWN.find(
-                    option => option.value === value,
-                );
-                if (selectedParameter) {
-                    updatedParams[index] = {
-                        ...updatedParams[index],
-                        name: selectedParameter.label,
-                        value: updatedParams[index]?.value || '',
-                    };
-                    setParamDisabled(prev => {
-                        const newStates = [...prev];
-                        newStates[index] = false;
-                        return newStates;
-                    });
-                } else {
-                    updatedParams[index] = { name: '', value: '' };
-                    setParamDisabled(prev => {
-                        const newStates = [...prev];
-                        newStates[index] = true;
-                        return newStates;
-                    });
-                }
-            } else if (keyValue === 'parameter_value') {
-                updatedParams[index] = {
-                    ...updatedParams[index],
-                    value,
-                };
-            }
-            const allParamsFilled = updatedParams.every(
-                param => param.value !== '',
-            );
-            if (
-                allParamsFilled &&
-                algorithm &&
-                entityType &&
-                entityTypeFields.length > 0
-            ) {
-                setConfirm(true);
-            } else {
-                setConfirm(false);
-            }
-
-            return updatedParams;
-        });
-    };
-
     useEffect(() => {
         const allParamsFilled = parameters.every(param => param.value !== '');
         if (
@@ -149,46 +92,7 @@ const AnalysisModal: FunctionComponent<Props> = ({ closeDialog, isOpen }) => {
         enabled: isOpen,
     });
 
-    const onChangeParameters = () => {
-        setParameterComponents(prevParams => [
-            ...prevParams,
-            `parameter_${prevParams.length}`,
-        ]);
-
-        setParameters(prevParams => [...prevParams, { name: '', value: '' }]);
-        setParamDisabled(prev => [...prev, true]);
-    };
-
-    const removeParameter = index => {
-        setParameterComponents(prevParams =>
-            prevParams.filter((_, i) => i !== index),
-        );
-
-        setParameters(prevParams => prevParams.filter((_, i) => i !== index));
-        setParamDisabled(prev => prev.filter((_, i) => i !== index));
-    };
-
-    const getFilteredOptions = useCallback(
-        index => {
-            const selectedValues = parameters.map(param => param.name);
-            const remainParams = LEVENSHTEIN_PARAMETERS_DROPDOWN.filter(
-                option =>
-                    !selectedValues.includes(option.label) ||
-                    parameters[index]?.name === option.label,
-            );
-
-            return remainParams;
-        },
-        [parameters],
-    );
-    const areOptionsAvailable = useMemo(
-        () =>
-            JSON.stringify(
-                LEVENSHTEIN_PARAMETERS_DROPDOWN.map(item => item.label),
-            ) === JSON.stringify(parameters.map(item => item.name)),
-        [parameters],
-    );
-
+    const isConfirm = algorithm && entityType && entityTypeFields.length > 0;
     return (
         <ConfirmCancelModal
             allowConfirm={confirm}
@@ -247,14 +151,10 @@ const AnalysisModal: FunctionComponent<Props> = ({ closeDialog, isOpen }) => {
                     />
                 </Grid>
                 <AnalysisModalParameters
-                    parameterComponents={parameterComponents}
                     parameters={parameters}
-                    onChangeParameters={onChangeParameters}
-                    handleParametersChange={handleParametersChange}
-                    getFilteredOptions={getFilteredOptions}
-                    paramDisabled={paramDisabled}
-                    removeParameter={removeParameter}
-                    areOptionsAvailable={areOptionsAvailable}
+                    setParameters={setParameters}
+                    setConfirm={setConfirm}
+                    isConfirm={isConfirm}
                 />
             </Grid>
         </ConfirmCancelModal>
